handle required preemption in generic x86 layer
We have been fudging attention from the dispatcher in the svm guest cpu run
code by setting kprunrun after a #VMEXIT.  It would be better to preempt only
if the userland runrun flag (or kprunrun) is set already, avoiding unecessary
switches.  It would better again to do this outside of the instruction
set-specific code so we'll do it in the generic x86 guest cpu run loop,
exiting back to userland if we should get off the CPU.

Note that the joyent vmx code already exits like this in an exit handler,
which I have marked with a comment.  We should test to see if we can remove
this from vmx, deferring to kvm_x86 to induce the exit.